src/requirements.txt
apscheduler
bcrypt
fastapi
psycopg2-binary
python-dotenv
python-jose
python-multipart
sqlalchemy
uvicorn